+++ packages/client/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Ensure mortality always encode to valid values.
+- ink!: Filter events by contract address for ink!v6
 
 ## 1.12.0 - 2025-05-21